+++ lib/cucumber/events/finished_testing.rb
@@ -1,9 +1,0 @@
-module Cucumber
-  module Events
-
-    # Event fired after aall test cases have finished executing
-    class FinishedTesting
-    end
-
-  end
-end