+++ lib/cucumber/events/test_run_finished.rb
@@ -1,0 +1,11 @@
+require 'cucumber/core/events'
+
+module Cucumber
+  module Events
+
+    # Event fired after all test cases have finished executing
+    class TestRunFinished < Core::Event.new
+    end
+
+  end
+end