+++ lib/cucumber/formatter/fail_fast.rb
@@ -7,8 +7,9 @@ module Cucumber
     class FailFast
 
       def initialize(configuration)
-        configuration.on_event :after_test_case do |event|
-          Cucumber.wants_to_quit = true unless event.result.ok?(configuration.strict?)
+        configuration.on_event :test_case_finished do |event|
+          test_case, result = *event.attributes
+          Cucumber.wants_to_quit = true unless result.ok?(configuration.strict?)
         end
       end